requirements.txt
tensorflow-gpu
tensorflow-datasets